+++ src/Spd.Resource.Applicants/ServiceExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Spd.Resource.Applicants.Application;
 using Spd.Resource.Applicants.ApplicationInvite;
-using Spd.Resource.Applicants.BulkHistory;
 using Spd.Utilities.Hosting;
 
 namespace Spd.Resource.Applicants
@@ -11,7 +10,6 @@ namespace Spd.Resource.Applicants
         public void ConfigureServices(ConfigurationServices configurationServices)
         {
             configurationServices.Services.AddTransient<IApplicationRepository, ApplicationRepository>();
-            configurationServices.Services.AddTransient<IBulkHistoryRepository, BulkHistoryRepository>();
             configurationServices.Services.AddTransient<IApplicationInviteRepository, ApplicationInviteRepository>();
         }
     }
